+++ extensions/disk-util/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [Initial Version] - {PR_MERGE_DATE}
 
-- Adds "Create Volume" command to create APFS encrypted and case-sensitive volumes.
+- Adds "Quick Create Volume" command to create APFS encrypted and case-sensitive volumes.
+- Adds "Create Volume" command to create APFS encrypted and case-sensitive volumes with extra features like symlinks (more features can be added later).